Refactored some things
