+++ android/app/src/main/kotlin/com/example/firstproject/MainActivity.kt
@@ -2,5 +2,4 @@ package com.example.firstproject
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
