+++ lib/answer.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
 class Answer extends StatelessWidget {
-  Function() selecteHandler;
-  String answerlist;
+  final Function() selectHandler;
+  final String answerList;
 
-  Answer(this.selecteHandler, this.answerlist, {super.key});
+  const Answer(this.selectHandler, this.answerList, {super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -12,9 +12,9 @@ class Answer extends StatelessWidget {
       width: double.infinity,
       child: ElevatedButton(
           onPressed: () {
-            selecteHandler();
+            selectHandler();
           },
-          child: Text(answerlist)),
+          child: Text(answerList)),
     );
   }
 }
