+++ lib/question.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class Question extends StatelessWidget {
   final String questionText;
 
-  Question(this.questionText);
+  const Question(this.questionText, {super.key});
 
   @override
   Widget build(BuildContext context) {
